Moving VAT form modal component to documents dir which is the only consumer
